examples/lightstep/requirements.txt
Flask
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-launcher
requests